+++ modules/image_io/image_tiff.h
@@ -8,7 +8,7 @@
 // getting file dimensions
 iPoint image_size_tiff(const std::string & file);
 
-Image image_load_tiff(const std::string & file, const int scale=1);
+Image image_load_tiff(const std::string & file, const double scale=1);
 
 //save the whole image
 void image_save_tiff(const Image & im, const std::string & file,
